+++ app/main/main_window.js
@@ -4,96 +4,84 @@ const {
     BrowserWindow,
 } = require('electron');
 
-const UserConfig = require('./user_config');
+const AppConfig = require('../common/app_config');
 
 
 module.exports = class MainWindow {
     constructor() {
         this.win = null;
-        this.config = {};
+        this.iconPath = null;
+        this.indexPath = null;
+        this.appConfig = new AppConfig();
     }
 
     setIcon(path) {
-        this.config.iconPath = path;
+        this.iconPath = path;
         return this;
     }
 
     setIndex(htmlFile) {
-        this.config.indexPath = htmlFile;
-        return this;
-    }
-
-    setUserConfig(userConfig) {
-        this.config.userConfig = userConfig;
+        this.indexPath = htmlFile;
         return this;
     }
 
     initWindow(devWindow) {
         let win;
-        const config = this.config.userConfig;
+        const windowSize = this.appConfig.get("windowSize");
 
-        if (!config.windowSize) config.windowSize = [500, 600]; //TODO: use default
         let winConfig = {
-            width: config.windowSize[0],
-            height: config.windowSize[1],
+            width: windowSize[0],
+            height: windowSize[1],
             minWidth: 365,
             minHeight: 300,
             fullscreenable: false,
             webgl: false,
-            icon: this.config.iconPath,
+            icon: this.iconPath,
             frame: true
         };
         if (devWindow) {
-            winConfig.width += config.devToolsSize;
+            winConfig.width += this.appConfig.get("devToolsSize");
         }
 
         win = new BrowserWindow(winConfig);
-        win.userConfig = config;
         win.setMenu(null);
 
-        win.loadURL(this.config.indexPath);
+        win.loadURL(this.indexPath);
 
         if (devWindow) win.webContents.openDevTools();
 
-        if (UserConfig.config.maximized === true) {
+        if (this.appConfig.get("maximized") === true) {
             win.maximize();
         }
 
         win.on('resize', () => {
-            if (UserConfig.config.maximized !== true) {
+            if (this.appConfig.get("maximized") !== true) {
                 let size = win.getSize();
-                if (devWindow) size[0] -= config.devToolsSize;
-                UserConfig.config.windowSize = size;
+                if (devWindow) size[0] -= this.appConfig.get("devToolsSize");
+                this.appConfig.set("windowSize", size);
             }
         });
 
         win.on('maximize', () => {
-            UserConfig.config.maximized = true;
+            this.appConfig.set("maximized", true);
         });
 
         win.on('unmaximize', () => {
-            UserConfig.config.maximized = false;
+            this.appConfig.set("maximized", false);
         });
 
         let first = true;
         win.on('close', (ev) => {
             if (first) {
                 ev.preventDefault();
-                UserConfig.saveConfig(() => {
-                    win.webContents.send('before-close');
-                    first = false;
-                });
+                win.webContents.send('before-close'); //TODO: remove?
+                first = false;
             }
         });
         win.on('closed', () => {
             win = null;
         });
 
-        this.clearConfig();
         return win;
-    }
-
-    clearConfig() {
-        this.config = {};
     }
 };