+++ app/renderer/app_status.js
@@ -1,10 +1,10 @@
 "use strict";
 
-const remote = require('electron').remote;
 const EventEmitter = require('events');
 
 const TaskConfig = require('./task_config');
 const version = require('../../package.json').version;
+const AppConfigStatus = require('./app_config_status');
 
 module.exports = {
     editMode: false,
@@ -14,7 +14,7 @@ module.exports = {
         this.editMode = !this.editMode;
         TaskConfig.saveConfig();
     },
-    config: remote.getCurrentWindow().userConfig, //TODO: improve how this works
+    config: new AppConfigStatus(),
     maxSuites: 6,
     maxTasksPerSuite: 8,
     totalTasks: 0,